메인페이지 댓글작성 부분 컴포넌트 분류 - Maincomment.js
